added barkeeper logic, added drink logic WIP

+++ models/ingredient.go
@@ -9,7 +9,7 @@ import (
 	Pump		Pump
 	PumpID		int
 	RecipeID	int
-	Parts 		uint16	`gorm:"notNull;"`
+	Parts 		uint	`gorm:"notNull;"`
   }
 
   func initIngredient() {
